+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Build script no longer uses grep to find package version
+
 ## [0.1.0] - 2025-11-20
 
 - Initial release